[RENAME] 외부 데이터 플랫폼 디렉토리 이동
